refactor(-): :hammer: move models carga dados
